Add LoadBalancer namespace and move classes
